# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pulse Surveys",
    "summary": "Weekly course feedback, confidential by construction.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/leadership/comparison-sets": {
      "get": {
        "tags": [
          "leadership"
        ],
        "summary": "Every comparison set this institution has defined",
        "description": "The institution's sets, in name order, each marked editable or not.\n\n**Institution-wide rather than this session's own** (ADR 0173). The reader\nstill comes from the session and from nowhere else: `editable` is decided\nagainst the `person` row this session was resolved to at the door, so no part\nof this request's own text decides which sets it may change.\n\n**Synchronous, and FastAPI runs it in a threadpool**, for the reason\n`app.api.instructor` gives: the session is synchronous (ADR 0013) and every\nstatement behind this blocks, so an `async` handler would take them on the\nevent loop.",
        "operationId": "read_sets_leadership_comparison_sets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "leadership"
        ],
        "summary": "Define a comparison set",
        "description": "Define a set for whoever this session is, or translate what the database refused.\n\n**The definer comes from the session and never from the body**, and it is\nwhat every later edit and delete of this set is scoped by.",
        "operationId": "define_set_leadership_comparison_sets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/leadership/comparison-sets/{set_id}": {
      "get": {
        "tags": [
          "leadership"
        ],
        "summary": "One comparison set, whoever defined it",
        "description": "One set and its membership.\n\n**A set id that is not a uuid is refused before this runs**, by FastAPI's own\nparsing of the path parameter \u2014 which is an answer about the shape of a\nrequest rather than about a set. ADR 0016 makes every key a uuid, so a\nmalformed value is not a set id at all and never reaches the lookup.",
        "operationId": "read_one_set_leadership_comparison_sets__set_id__get",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Set Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "leadership"
        ],
        "summary": "Replace the definition of a set I defined",
        "description": "Replace a set's whole definition, for the leader who defined it.\n\nThe three refusals in order: an id nothing defined is a 404, another leader's\nset is a 403, and a definition the database will not store is one of the five\ntranslated sentences.",
        "operationId": "replace_set_leadership_comparison_sets__set_id__put",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Set Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "leadership"
        ],
        "summary": "Delete a set I defined",
        "description": "Delete a set, taking its membership rows and nothing else (ADR 0164).\n\n204 and no body, because there is nothing left to describe. A course named by\nthe set is untouched and stays undeletable while any set names it, which is\nthe opposite direction of the same rule.",
        "operationId": "remove_set_leadership_comparison_sets__set_id__delete",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Set Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/leadership/comparison-sets/{set_id}/preview": {
      "get": {
        "tags": [
          "leadership"
        ],
        "summary": "How many courses and sections a set reaches",
        "description": "Two counts and nothing else \u2014 see the module docstring on what that costs and why.\n\nAnswered for every set, not only for this session's own: a count is what\nmakes an institution-wide list usable, and it says nothing about anybody.",
        "operationId": "read_preview_leadership_comparison_sets__set_id__preview_get",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Set Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SetDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "length_weeks": {
            "type": "integer",
            "title": "Length Weeks"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "member_count": {
            "type": "integer",
            "title": "Member Count"
          },
          "editable": {
            "type": "boolean",
            "title": "Editable"
          },
          "member_course_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Member Course Ids"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "length_weeks",
          "level",
          "member_count",
          "editable",
          "member_course_ids",
          "created_at",
          "updated_at"
        ],
        "title": "SetDetail",
        "description": "One set on its own: the summary, its membership, and when it was written.\n\nThe two instants are the whole record that a set was created and edited (ADR\n0174): this ticket writes no `audit_log` row, and these columns are what is\nrecorded instead."
      },
      "SetList": {
        "properties": {
          "sets": {
            "items": {
              "$ref": "#/components/schemas/SetSummary"
            },
            "type": "array",
            "title": "Sets"
          }
        },
        "type": "object",
        "required": [
          "sets"
        ],
        "title": "SetList",
        "description": "Every set in the institution, in name order."
      },
...
      "SetPreview": {
        "properties": {
          "member_count": {
            "type": "integer",
            "title": "Member Count"
          },
          "section_count": {
            "type": "integer",
            "title": "Section Count"
          }
        },
        "type": "object",
        "required": [
          "member_count",
          "section_count"
        ],
        "title": "SetPreview",
        "description": "What a set reaches, as two counts and nothing else. See the module docstring."
      },
      "SetSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "length_weeks": {
            "type": "integer",
            "title": "Length Weeks"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "member_count": {
            "type": "integer",
            "title": "Member Count"
          },
          "editable": {
            "type": "boolean",
            "title": "Editable"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "length_weeks",
          "level",
          "member_count",
          "editable"
        ],
        "title": "SetSummary",
        "description": "One set as it appears in the institution-wide list.\n\n`member_count` is the number of member courses, not of resolved sections \u2014\nthe preview answers that, and the two are different numbers for every set\nwhose courses run more than one section."
      },
      "SetWrite": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "length_weeks": {
            "type": "integer",
            "title": "Length Weeks"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "member_course_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Member Course Ids"
          }
        },
        "type": "object",
        "required": [
          "name",
          "length_weeks",
          "level",
          "member_course_ids"
        ],
        "title": "SetWrite",
        "description": "What a client sends to create or edit a comparison set.\n\nThe same body on `POST` and on `PUT`, because an edit replaces a set's whole\ndefinition \u2014 its name, its declared pair, and its membership \u2014 rather than\npatching part of one. A partial edit would make \"what is this set\" a question\nabout which fields arrived.\n\n`length_weeks` and `level` are deliberately unconstrained here. See the module\ndocstring: the database holds both rules and this route translates what it\nrefuses.\n\n**The name's surrounding spaces are stripped, and a name empty after that\nis refused here** (E5-14), so \" Nursing\" and \"Nursing\" are one name to the\nunique constraint, and a name of only spaces is answered with the framework's\nstandard 422 before any statement runs. That is the one rule this model\nholds itself, and it is the ruling's choice over a new refusal sentence: the\ntable's `name_is_not_blank` check stays underneath it for every other writer,\nbut the translator names no sentence for it, so a blank name reaching the\ntable from this route would be a 500."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}